requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bitsandbytes==0.45.4
bleach==6.2.0
boto3==1.37.16
botocore==1.37.16
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
flash_attn==2.7.4.post1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.71.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.0
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.5.2
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.0.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
latex2mathml==3.77.0
lightning==2.5.1
lightning-cloud==0.5.70
lightning-utilities==0.14.2
lightning_sdk==0.2.4
litdata==0.2.32
litserve==0.2.6
llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@09e5840d5589ad2d6a8656c0a60f21ae134b3309
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==4.23.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.23.1
rsa==4.9
ruff==0.11.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.24.1
setproctitle==1.3.5
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simple-term-menu==1.6.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.6.0
torchmetrics==1.3.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.0
triton==3.2.0
trl==0.16.1
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tyro==0.8.14
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
wandb==0.19.8
watchfiles==1.0.4
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.20.0
